Исправлены все ошибки проверок.

+++ core/src/main/kotlin/ru/kchvilyov/taskmanager/core/presentation/TaskIntent.kt
@@ -3,6 +3,6 @@ package ru.kchvilyov.taskmanager.core.presentation
 import ru.kchvilyov.taskmanager.core.domain.Task
 
 sealed interface TaskIntent {
-    data class ToggleTask(val taskId: Int) : TaskIntent
+    data class ToggleTask(val taskId: Long) : TaskIntent
     data class AddTask(val task: Task) : TaskIntent
 }
